requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
pillow==10.1.0
pydantic==2.5.0
python-dotenv==1.0.0
langchain-groq==0.0.1
transformers==4.36.0
torch==2.1.0
torchvision==0.16.0
numpy==1.24.3
opencv-python==4.8.1.78
matplotlib==3.8.2
captum==0.6.0
lime==0.2.0.1
ultralytics==8.0.206
streamlit==1.28.1